requirements.txt
csv
bs4
json
numpy
talib
pandas
urllib
pydash
httplib2
slackclient
configparser
oauth2client
urllib.request
apiclient.http
googleapiclient.discovery
google.oauth2.credentials
google_auth_oauthlib.flow


